Tryint to move to single run file
