--- Image_data/Coco_HPCrun
@@ -1,6 +1,0 @@
-#!/bin/bash
-echo "Makeing Output Directory"
-mkdir -p Coco_output 
-
-echo "Running Job Array"
-sbatch Coco.sb
