--- Image_data/Sky.sb
@@ -1,55 +1,0 @@
-#!/bin/bash --login
-#SBATCH -c 1
-#SBATCH -n 1
-#SBATCH --time=4:00:00
-#SBATCH --mem=2gb
-#SBATCH --mail-type=END
-#SBATCH --array 0-59
-#SBATCH --output "./Sky_output/%a_%A_slurm.output"
-
-#Log a copy of this script in the output file
-cat $0
-
-echo "Job running on ${HOSTNAME}"
-
-export PYTHONPATH=${PYTHONPATH}:${SLURM_SUBMIT_DIR}:${SLURM_SUBMIT_DIR}/../
-
-out_folder=${TMPDIR}/${SLURM_ARRAY_TASK_ID}/
-echo Output sent to $out_folder 
-mkdir -p $out_folder 
-
-tarfile=${SLURM_ARRAY_TASK_ID}_jobfiles.tgz
-if [ -f ${out_folder}/${tarfile} ]
-then
-	echo "Loading Checkpoint"
-	cp ${out_folder}/${tarfile} ${TMPDIR}
-	cd ${TMPDIR}
-	tar -xzvf $tarfile
-	cd ${SLURM_SUBMIT_DIR}
-	
-fi
-
-srun --unbuffered python SkyRun.py -o $out_folder -i ${SLURM_ARRAY_TASK_ID} -g 100 -p 500 -s ${SLURM_ARRAY_TASK_ID} > ${out_folder}/search.out 
-
-cp CocoReport.ipynb ${out_folder}
-
-cd ${out_folder}
-srun time jupyter nbconvert --no-input --allow-errors --execute --to html CocoReport.ipynb
-
-cp CocoReport.html ${SLURM_SUBMIT_DIR}/Sky_output/${SLURM_ARRAY_TASK_ID}_Report.html 
-
-#zip up this folder
-cd ${TMPDIR}
-tar -czvf ${SLURM_SUBMIT_DIR}/Sky_output/${tarfile} ./${SLURM_ARRAY_TASK_ID}/ 
-
-#House Cleaning
-cd ${SLURM_SUBMIT_DIR} 
-rm -rf ${TMPDIR}/*
-
-#Job Statistics
-module load powertools
-js -j $SLURM_JOB_ID
-
-echo "Done"
-
-##########################
