--- Image_data/Sky_HPCrun
@@ -1,6 +1,0 @@
-#!/bin/bash
-echo "Makeing Output Directory"
-mkdir -p Sky_output 
-
-echo "Running Job Array"
-sbatch Sky.sb
